requirements.txt
flask
slack_bolt
flask-bcrypt
requests
pyairtable
arrow
gunicorn
python-dotenv
uuid
importlib-resources
importlib-metadata
zipp
svcs
psycopg[binary]
pydantic
bs4
openai
boto3
rich